Refactor controller endpoints and unify count type to Integer
Added leading slashes to @RequestMapping paths in controllers and templates for consistency. Changed the return type of selectCount from Long to Integer across Mapper XML, VO, and test code. Updated service and controller bean names for clarity. These changes improve endpoint clarity and unify pagination count handling.

+++ sql2codeInitDemo/src/main/java/com/slilio/sql2codeInitDemo/entity/vo/PaginationResultVO.java
@@ -34,11 +34,11 @@ public class PaginationResultVO<T> {
 
   public PaginationResultVO() {}
 
-  public Long getTotalCount() {
+  public Integer getTotalCount() {
     return totalCount;
   }
 
-  public void setTotalCount(Long totalCount) {
+  public void setTotalCount(Integer totalCount) {
     this.totalCount = totalCount;
   }
